Rename the "animation" module to the "ui" module
