+++ settings.gradle
@@ -20,4 +20,4 @@ include ':android-test-common',
         ':subscribing-sdk-java',
         ':subscribing-example-app',
         ':subscribing-java-testing',
-        ':animation'
+        ':ui-sdk'
